Renamed vector db modules to vector store.
